requirements.txt
streamlit-0.79.0-py2.py3-none-any.whl